backend/requirements.txt
# FastAPI and web server
fastapi==0.115.0
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
pymysql==1.1.1

# Configuration
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# File uploads and processing
aiofiles==24.1.0

# Data processing and AI/ML
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
filterpy==0.1.8

# MinIO (optional)
minio==7.2.8

# Background tasks (optional)
celery==5.4.0
redis==5.2.1

# Development and testing
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1

# Code formatting and linting
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.1
